requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-core
langchain-community
langserve
langchain-openai
langchain-ollama
python-dotenv
streamlit
setuptools
wheel
